Update README.md
Various methods were used to filter the data and visualize the driver's position in real-time. The dataset contains information about the drivers and their respective lap times, allowing for a detailed analysis of each driver's performance. This project is an excellent resource for anyone interested in data analysis or motorsports, providing a detailed insight into drivers' positions during the race.